T33.P2: bf16 support — P1.5 / P1.6 / miles integration shim all on bf16
mlir-ascend tilelang dtype = 'bfloat16' is accepted by the codegen
with zero kernel changes. Validation:

  P1.5 lighting_indexer_fwd  bf16: max abs err 0.000000 (matches CPU ref)
  P1.6 lighting_indexer_bwd  bf16:
     dQ err 9e-5  (was 1e-5 in fp16)
     dW err 0     (exact)
     dKV err 4e-5
  miles V4IndexerFunctionNPU bf16:
     dq err 2.3e-3  (within fp16/bf16 quantization tolerance)
     dw err 0       (exact)

bf16 widens the dynamic range and trades mantissa precision; the
slight err increase is expected. Still well within training tolerance
(typically << 1e-2). miles uses bf16 in its V4 wrapper, so this is
the production-correct dtype.

Now all P1 production-passing kernels are bf16:
  * sparse_mla_fwd (P1.3): fp16 (V4 spec uses fp16 here too)
  * lighting_indexer_fwd (P1.5): bf16 ✓
  * lighting_indexer_bwd (P1.6): bf16 ✓
  * miles integration shim (P2.miles_integration_smoke): bf16 ✓

+++ examples/deepseek_v4/example_lighting_indexer_bwd_kernel.py
@@ -54,7 +54,7 @@ def lighting_indexer_bwd(
     Returns dIndexQ via out_idx=[-2]; dWeights and dIndexK are written
     in-place into caller-provided tensors.
     """
-    dtype = "float16"
+    dtype = "bfloat16"  # miles uses bf16
     accum_dtype = "float32"
     idx_dtype = "int32"
 
@@ -271,8 +271,8 @@ def _smoke_bwd():
 
     # Inputs
     torch.manual_seed(0)
-    q = torch.randn(SEQ, H, D, dtype=torch.float16, device="npu") * 0.1
-    kv = torch.randn(SKV, D, dtype=torch.float16, device="npu") * 0.1
+    q = torch.randn(SEQ, H, D, dtype=torch.bfloat16, device="npu") * 0.1
+    kv = torch.randn(SKV, D, dtype=torch.bfloat16, device="npu") * 0.1
     w = torch.randn(SEQ, H, dtype=torch.float32, device="npu") * 0.5
     topk_idx = torch.zeros(SEQ, K, dtype=torch.int32, device="npu")
     for s in range(SEQ):
